add 19.11.2022

+++ figure_relations/triangle.h
@@ -6,7 +6,7 @@
 class Triangle: public Figures
 {
 	protected:
-		unsigned a,b,c;
+	float a,b,c;
 
 	public:
 		/// Конструктор по умолчанию
